requirements.txt
pyyaml>=6.0
pytest>=7.0 